Add message unpack stream
